backend/requirements.txt
fastapi
uvicorn
pydantic
PyJWT
email-validator
pytest
